requirements.txt
fastapi==0.85.0
uvicorn==0.18.3
pydantic==1.10.2
opentelemetry-distro==0.33b0
opentelemetry-exporter-otlp-proto-grpc==1.12.0
opentelemetry-instrumentation-fastapi==0.33b0
